WIP: Currently dealing with memory leak
